services/deepfake/requirements.txt
flask==2.3.2
flask-cors==3.0.10
numpy==1.24.3
opencv-python==4.7.0.72
onnxruntime-gpu==1.15.1
pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
efficientnet-pytorch==0.7.1
